+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - Ethernet Network
    - FC Network
    - FCoE Network
+   - Logical Interconnect
    - Logical Interconnect Group
    - Logical Switch Group
    - Network Set